+++ sv-linx/src/main/java/com/hartwig/hmftools/linx/annotators/KataegisData.java
@@ -3,12 +3,12 @@ package com.hartwig.hmftools.linx.annotators;
 public class KataegisData
 {
     public final String Chromosome;
-    public final long PosStart;
-    public final long PosEnd;
+    public final int PosStart;
+    public final int PosEnd;
     public final String Id;
     public final int SnvCount;
 
-    public KataegisData(final String chromosome, long posStart, long posEnd, final String id, int snvCount)
+    public KataegisData(final String chromosome, int posStart, int posEnd, final String id, int snvCount)
     {
         Chromosome = chromosome;
         PosStart = posStart;
